requirements.txt
pyyaml
selenium
pyperclip